+++ syncrclone/__init__.py
@@ -1,5 +1,5 @@
-__version__ = "20221229.0.BETA"
-LASTRCLONE = "1.61.0"  # This is the last version I tested with. Does *NOT* mean it won't work further.
+__version__ = "20221230.0.BETA"
+LASTRCLONE = "1.61.1"  # This is the last version I tested with. Does *NOT* mean it won't work further.
 
 import time
 import io